+++ src/Config/Option/QrCode/DefaultMarginConfigOption.php
@@ -2,29 +2,28 @@
 
 declare(strict_types=1);
 
-namespace Shlinkio\Shlink\Installer\Config\Option\UrlShortener;
+namespace Shlinkio\Shlink\Installer\Config\Option\QrCode;
 
 use Shlinkio\Shlink\Config\Collection\PathCollection;
 use Shlinkio\Shlink\Installer\Config\Option\BaseConfigOption;
 use Shlinkio\Shlink\Installer\Config\Util\ConfigOptionsValidatorsTrait;
 use Symfony\Component\Console\Style\StyleInterface;
 
-class ShortCodeLengthOption extends BaseConfigOption
+class DefaultMarginConfigOption extends BaseConfigOption
 {
     use ConfigOptionsValidatorsTrait;
 
     public function getConfigPath(): array
     {
-        return ['url_shortener', 'default_short_codes_length'];
+        return ['qr_codes', 'margin'];
     }
 
     public function ask(StyleInterface $io, PathCollection $currentOptions): int
     {
         return $io->ask(
-            'What is the default length you want generated short codes to have? (You will still be able to override '
-            . 'this on every created short URL)',
-            '5',
-            fn ($value) => $this->validateNumberGreaterThan($value, 4),
+            'What\'s the default margin, in pixels, you want generated QR codes to have',
+            '0',
+            fn (mixed $value) => $this->validateNumberGreaterThan($value, 0),
         );
     }
 }